fix: missing forward declaration and remaining narrowing conversions
- G1G2G3Native.cpp: add forward declaration of resolveWithTimeout before dnsPollution
  Fixes: error: use of undeclared identifier 'resolveWithTimeout'
- G1G2G3Native.cpp: static_cast<int> on qMax/elapsed (lines 171,502-505,2294)
- NetworkProbe.cpp: static_cast<int> on timer.elapsed, auto for batchSize
- G5WebsiteUrl.cpp: static_cast<int> on QString::length and qMax
- PingParser.cpp: static_cast<int> on result.hops.size()

+++ src/engine/diagnostic/G5WebsiteUrl.cpp
@@ -356,10 +356,10 @@ DiagnosticResult sslCertificate(const QString& target) {
         QString::number(cert.daysLeft), QString::number(cert.subjectAltNames.size()),
         cert.thumbprint.left(40)};
 
-    int nw = QStringLiteral("thumbprint").length();  // longest label
-    for (const auto& s : names) nw = qMax(nw, s.length());
+    int nw = static_cast<int>(QStringLiteral("thumbprint").length());
+    for (const auto& s : names) nw = qMax(nw, static_cast<int>(s.length()));
     int vw = 0;
-    for (const auto& s : vals) vw = qMax(vw, s.length());
+    for (const auto& s : vals) vw = qMax(vw, static_cast<int>(s.length()));
 
     lines.append(QStringLiteral("*  %1  %2")
         .arg(QStringLiteral("Property"), -nw)
